+++ test/main.js
@@ -79,8 +79,9 @@ $(document)
 				dataProvider : {
 					// you can pre-populate 'data' or the refresh() function will be called when you instantiate the ComboBox
 					//data : undefined,
-					refresh : function(){
-						this.data = foods;
+					refresh : function( dataProvider, resolve, reject ){
+						dataProvider.data = foods;
+						resolve();
 					}
 				}
 			}